+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/shipping/InOutToTransportationOrderService.java
@@ -25,7 +25,6 @@ package de.metas.handlingunits.shipping;
 import com.google.common.collect.ImmutableList;
 import de.metas.handlingunits.IHUShipperTransportationBL;
 import de.metas.handlingunits.IHandlingUnitsBL;
-import de.metas.handlingunits.impl.CreatePackagesForInOutRequest;
 import de.metas.handlingunits.inout.IHUInOutDAO;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.model.I_M_InOut;
